2005-08-15  Martin Baulig  <martin@ximian.com>
	* backends/remoting/: Update to support multiple connections and
	also correctly exit when we're done.


svn path=/trunk/debugger/; revision=48396

+++ backends/server/remoting.c
@@ -70,7 +70,8 @@ void
 mono_debugger_remoting_kill (int pid, int fd)
 {
 	close (fd);
-	kill (pid, SIGKILL);
+	if (pid)
+		kill (pid, SIGKILL);
 }
 
 int
@@ -99,7 +100,7 @@ mono_debugger_remoting_poll (int fd, PollFunc func)
 			break;
 		}
 
-		if (fds [0].revents & (POLLHUP | POLLERR))
+		if (fds [0].revents & (POLLHUP | POLLERR | POLLNVAL))
 			break;
 
 		func ();
